refractor: classes to best jpa practice

+++ src/main/java/hei/school/demo/endpoint/rest/controller/dto/BookCopyCreationDto.java
@@ -1,5 +1,3 @@
 package hei.school.demo.endpoint.rest.controller.dto;
 
-import java.math.BigDecimal;
-
-public record BookCopyCreationDto(String bookId, String format, BigDecimal sellingPrice) {}
+public record BookCopyCreationDto(String bookId, String format, Double sellingPrice) {}
